move errorboundary into components folder
