+++ src/components/ErrorBoundary/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { connect } from 'react-redux';
-import AppError from '../AppError';
+import AppError from './AppError';
 import reportError from '../../state/reportError';
 import { withRouter } from 'react-router';
